Update README.md
I am excited to embark on a Python course inspired by the teachings of Sir Ammar from Codanics Python class. As I delve into this course, I'll be sharing the valuable insights and knowledge gained with you. Stay tuned for updates and lessons from this enriching learning journey. Thank you for joining me on this adventure!