Update README.md
Lesson 1: Variables

This lesson introduces the concept of variables in Python programming. It covers the basics of numeric and string variables, rules for naming variables, identifying variable types, and deleting variables using the `del` function. By the end of the lesson, learners will have a solid understanding of how to effectively use variables in Python programs.